Reorganize view-related code.
